+++ core/src/main/scala/at/logic/gapt/proofs/ceres/StandardClauseSet.scala
@@ -8,10 +8,6 @@ package at.logic.gapt.proofs.ceres
 import at.logic.gapt.proofs.{ HOLClause, HOLSequent, Sequent, SetSequent }
 import at.logic.gapt.expr._
 import at.logic.gapt.proofs.lk.LKProof
-import at.logic.gapt.utils.Logger
-
-import scala.annotation.tailrec
-import scala.util.control.TailCalls._
 
 /**
  * Calculates the characteristic clause set