+++ examples/reforest.scala
@@ -1,8 +1,9 @@
 package at.logic.gapt.examples
 
-import at.logic.gapt.cutintro.{ ReforestMethod, CutIntroduction }
+import at.logic.gapt.cutintro.{ CutIntroduction, ReforestMethod }
 import at.logic.gapt.grammars.reforest.Reforest
 import at.logic.gapt.proofs.expansion.InstanceTermEncoding
+import at.logic.gapt.utils.verbose
 
 object ReforestDemo extends Script {
   val ( lang, _ ) = InstanceTermEncoding(
@@ -12,8 +13,9 @@ object ReforestDemo extends Script {
   println( grammar.toRecursionScheme )
   //  println( grammar.toVTRATG )
 
-  CutIntroduction.makeVerbose()
-  CutIntroduction(
-    SumExampleProof( 16 ),
-    method = ReforestMethod )
+  verbose {
+    CutIntroduction(
+      SumExampleProof( 16 ),
+      method = ReforestMethod )
+  }
 }